Move file: merge login&register
