--- src/modules/common/hooks/useLogin.ts
@@ -1,9 +1,0 @@
-import { login, type LoginBodyParam } from "@/modules/services/auth";
-import { useMutation } from "@tanstack/react-query";
-
-export const useLogin = () => {
-  return useMutation({
-    mutationFn: ({ username, password }: LoginBodyParam) =>
-      login({ username, password }),
-  });
-};
